Placeholder: move resource config store init logic
